BENCH-375: Animate quantity button & modals
