+++ src/components/ProductCard/ProductCard.tsx
@@ -1,7 +1,7 @@
 import PlaceHolderImage from 'assets/burger_transparent.png';
 import { BadgeIcon } from 'components/Icons/BadgeIcon';
 import { Button } from 'components/Button/Button';
-import { MenuProductModal } from 'components/MenuProductModal/MenuProductModal';
+import { MenuProductModal } from 'components/Modals/MenuProductModal';
 import { ProductDto } from 'dtos/ProductDto';
 import { ReactElement, useState } from 'react';
 import { motion } from 'framer-motion';
